requirements.txt
sh
PyYAML
pathlib
